[caffe2][nomnigraph] Move tests to proper gtest suite
