+++ caffe2/CMakeLists.txt
@@ -179,7 +179,7 @@ if (BUILD_TEST)
     # libraries even if they are not part of the public dependency libs. As a
     # result, we will explicitly link the test against the Caffe2 dependency
     # libs.
-    target_link_libraries(${test_name} ${Caffe2_MAIN_LIBS} gtest_main)
+    target_link_libraries(${test_name} ${Caffe2_DEPENDENCY_LIBS} ${Caffe2_MAIN_LIBS} gtest_main)
     if (USE_CUDA)
       target_link_libraries(${test_name} ${Caffe2_CUDA_DEPENDENCY_LIBS})
     endif()
