+++ caffe2/core/nomnigraph/include/nomnigraph/Converters/Dot.h
@@ -5,6 +5,7 @@
 #include "nomnigraph/Support/Casting.h"
 
 #include <functional>
+#include <map>
 #include <iostream>
 #include <sstream>
