+++ caffe2/core/nomnigraph/tests/binary_match_test.cc
@@ -1,82 +1,11 @@
 #include <gtest/gtest.h>
 #include <set>
 
+#include "test_util.h"
+
 #include "nomnigraph/Converters/Dot.h"
 #include "nomnigraph/Graph/Algorithms.h"
 #include "nomnigraph/Graph/Graph.h"
-
-/** Our test graph looks like this:
- *           +-------+
- *           | entry |
- *           +-------+
- *             |
- *             |
- *             v
- *           +-------+
- *           |   1   |
- *           +-------+
- *             |
- *             |
- *             v
- * +---+     +-------+
- * | 4 | <-- |   2   |
- * +---+     +-------+
- *   |         |
- *   |         |
- *   |         v
- *   |       +-------+
- *   |       |   3   |
- *   |       +-------+
- *   |         |
- *   |         |
- *   |         v
- *   |       +-------+
- *   +-----> |   6   |
- *           +-------+
- *             |
- *             |
- *             v
- * +---+     +-------+
- * | 5 | --> |   7   |
- * +---+     +-------+
- *             |
- *             |
- *             v
- *           +-------+
- *           | exit  |
- *           +-------+
- *
- * Here is the code used to generate the dot file for it:
- *
- *  auto str = nom::converters::convertToDotString(&graph,
- *    [](nom::Graph<std::string>::NodeRef node) {
- *      std::map<std::string, std::string> labelMap;
- *      labelMap["label"] = node->data();
- *      return labelMap;
- *    });
- */
-nom::Graph<std::string> createGraph() {
-  nom::Graph<std::string> graph;
-  auto entry = graph.createNode(std::string("entry"));
-  auto n1 = graph.createNode(std::string("1"));
-  auto n2 = graph.createNode(std::string("2"));
-  auto n3 = graph.createNode(std::string("3"));
-  auto n4 = graph.createNode(std::string("4"));
-  auto n5 = graph.createNode(std::string("5"));
-  auto n6 = graph.createNode(std::string("6"));
-  auto n7 = graph.createNode(std::string("7"));
-  auto exit = graph.createNode(std::string("exit"));
-  graph.createEdge(entry, n1);
-  graph.createEdge(n1, n2);
-  graph.createEdge(n2, n3);
-  graph.createEdge(n2, n4);
-  graph.createEdge(n3, n6);
-  graph.createEdge(n4, n6);
-  graph.createEdge(n6, n7);
-  graph.createEdge(n5, n7);
-  graph.createEdge(n7, exit);
-  return graph;
-}
 
 TEST(BinaryMatch, NoMatch) {
   auto graph = createGraph();
